+++ be/src/olap/comparison_predicate.cpp
@@ -111,17 +111,17 @@ COMPARISON_PRED_EVALUATE(GreaterEqualPredicate, >=)
 
 #define COMPARISON_PRED_COLUMN_BLOCK_EVALUATE(CLASS, OP) \
     template<class type> \
-    void CLASS<type>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const { \
+    void CLASS<type>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const { \
         uint16_t new_size = 0; \
         if (block->is_nullable()) { \
-            for (int i = 0; i < *size; ++i) { \
+            for (uint16_t i = 0; i < *size; ++i) { \
                 uint16_t idx = sel[i]; \
                 sel[new_size] = idx; \
                 const type* cell_value = reinterpret_cast<const type*>(block->cell(idx).cell_ptr()); \
                 new_size += (!block->cell(idx).is_null() && (*cell_value OP _value)); \
             } \
         } else { \
-            for (int i = 0; i < *size; ++i) { \
+            for (uint16_t i = 0; i < *size; ++i) { \
                 uint16_t idx = sel[i]; \
                 sel[new_size] = idx; \
                 const type* cell_value = reinterpret_cast<const type*>(block->cell(idx).cell_ptr()); \
@@ -179,17 +179,17 @@ COMPARISON_PRED_EVALUATE_DECLARATION(GreaterPredicate)
 COMPARISON_PRED_EVALUATE_DECLARATION(GreaterEqualPredicate)
 
 #define COMPARISON_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(CLASS) \
-    template void CLASS<int8_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int16_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int32_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int128_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<float>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<double>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<decimal12_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<StringValue>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<uint24_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<uint64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
+    template void CLASS<int8_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int16_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int32_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int128_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<float>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<double>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<decimal12_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<StringValue>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<uint24_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<uint64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
 
 COMPARISON_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(EqualPredicate)
 COMPARISON_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(NotEqualPredicate)
