+++ be/src/olap/in_list_predicate.cpp
@@ -86,17 +86,17 @@ IN_LIST_PRED_EVALUATE(NotInListPredicate, ==)
 
 #define IN_LIST_PRED_COLUMN_BLOCK_EVALUATE(CLASS, OP) \
     template<class type> \
-    void CLASS<type>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const { \
+    void CLASS<type>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const { \
         uint16_t new_size = 0; \
         if (block->is_nullable()) { \
-            for (int i = 0; i < *size; ++i) { \
+            for (uint16_t i = 0; i < *size; ++i) { \
                 uint16_t idx = sel[i]; \
                 sel[new_size] = idx; \
                 const type* cell_value = reinterpret_cast<const type*>(block->cell(idx).cell_ptr()); \
                 new_size += (!block->cell(idx).is_null() && _values.find(*cell_value) OP _values.end()); \
             } \
         } else { \
-            for (int i = 0; i < *size; ++i) { \
+            for (uint16_t i = 0; i < *size; ++i) { \
                 uint16_t idx = sel[i]; \
                 sel[new_size] = idx; \
                 const type* cell_value = reinterpret_cast<const type*>(block->cell(idx).cell_ptr()); \
@@ -142,17 +142,17 @@ IN_LIST_PRED_EVALUATE_DECLARATION(InListPredicate)
 IN_LIST_PRED_EVALUATE_DECLARATION(NotInListPredicate)
 
 #define IN_LIST_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(CLASS) \
-    template void CLASS<int8_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int16_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int32_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<int128_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<float>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<double>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<decimal12_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<StringValue>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<uint24_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
-    template void CLASS<uint64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint32_t* size) const; \
+    template void CLASS<int8_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int16_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int32_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<int128_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<float>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<double>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<decimal12_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<StringValue>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<uint24_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
+    template void CLASS<uint64_t>::evaluate(ColumnBlock* block, uint16_t* sel, uint16_t* size) const; \
 
 IN_LIST_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(InListPredicate)
 IN_LIST_PRED_COLUMN_BLOCK_EVALUATE_DECLARATION(NotInListPredicate)
